Removed service and added Test case to verify dialog
