+++ app/controllers/application.js
@@ -4,21 +4,13 @@ import { MAIN_SITE_URL } from '../constants/url';
 import { action } from '@ember/object';
 import { GITHUB_URL } from '../constants/url';
 import ENV from 'website-my/config/environment';
-import { checkDeviceType } from '../utils/checkDeviceType';
 
 export default class ApplicationController extends Controller {
   @service router;
-  @service rdsdialogState;
-
   GITHUB_URL = GITHUB_URL;
   BASE_API_URL = ENV.BASE_API_URL;
   get canShowNavBar() {
     return this.router.currentRouteName != 'signup';
-  }
-
-  constructor() {
-    super(...arguments);
-    this.rdsdialogState.updateData(checkDeviceType());
   }
 
   @action async signOutHandler() {
