+++ tests/integration/components/mobile-dialog-test.js
@@ -6,21 +6,10 @@ import { hbs } from 'ember-cli-htmlbars';
 module('Integration | Component | mobile-dialog', function (hooks) {
   setupRenderingTest(hooks);
 
-  test('it renders', async function (assert) {
-    // Set any properties with this.set('myProperty', 'value');
-    // Handle any actions with this.set('myAction', function(val) { ... });
+  test('Mobile-Dialog does not renders', async function (assert) {
 
-    await render(hbs`<MobileDialog />`);
+    await render(hbs`<MobileDialog/>`); 
 
-    assert.equal(this.element.textContent.trim(), '');
-
-    // Template block usage:
-    await render(hbs`
-      <MobileDialog>
-        template block text
-      </MobileDialog>
-    `);
-
-    assert.equal(this.element.textContent.trim(), 'template block text');
+    assert.dom('.appDialog').doesNotExist()
   });
 });
