--- tests/unit/services/rdsdialog-state-test.js
@@ -1,12 +1,0 @@
-import { module, test } from 'qunit';
-import { setupTest } from 'ember-qunit';
-
-module('Unit | Service | RDSDialog-state', function (hooks) {
-  setupTest(hooks);
-
-  // TODO: Replace this with your real tests.
-  test('it exists', function (assert) {
-    let service = this.owner.lookup('service:rdsdialog-state');
-    assert.ok(service);
-  });
-});
